refactor(advanced): 아이콘 컴포넌트 및 SVG 프로바이더를 FSD 아키텍처에 맞게 모듈 분리
- 애플리케이션 진입점 App 컴포넌트 app으로 이동
- 전역 SVG 프로바이더 app/providers으로 이동
- 아이콘 컴포넌트 및 아이콘명은 shared/icon으로 이동
