+++ src/advanced/app/index.ts
@@ -1,0 +1,2 @@
+export { App } from './App';
+export { GlobalSVGProvider } from './providers/GlobalSVGProvider';
