--- src/advanced/components/icons/index.ts
@@ -1,2 +1,0 @@
-export * from './Icon';
-export * from './GlobalSVGProvider';
