+++ src/advanced/components/ui/CartHeader.tsx
@@ -1,6 +1,6 @@
 import type { CartItem } from '../../../types';
 import { calculateItemTotalQuantity } from '../../models/cart';
-import { Icon } from '../icons';
+import { Icon } from '../../shared/icon';
 
 interface CartHeaderProps {
   searchTerm: string;
