+++ src/advanced/main.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './App.tsx';
-import { GlobalSVGProvider } from './components/icons/GlobalSVGProvider.tsx';
+import { App, GlobalSVGProvider } from './app';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
