+++ src/advanced/shared/icon/index.ts
@@ -1,0 +1,2 @@
+export { Icon } from './ui/Icon';
+export { iconNames } from './consts/iconNames';
